requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.5.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bitarray==3.0.0
bitstring==4.3.0
bleach==6.2.0
cachetools==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Django==5.2.7
django-filter==25.2
djangorestframework==3.15.2
docx==0.2.4
ecdsa==0.19.0
esptool==4.8.1
et_xmlfile==2.0.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
fire==0.7.1
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
fsspec==2024.2.0
gast==0.6.0
google-auth==2.41.1
google-genai==1.42.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
intelhex==2.3.0
ipykernel==6.29.5
ipython==8.32.0
isoduration==20.11.0
jdcal==1.4.1
jedi==0.19.2
Jinja2==3.1.3
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
lxml==6.0.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.9
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook_shim==0.2.4
numpy==2.2.6
opencv-python==4.10.0.84
opencv-python-headless==4.12.0.88
openpyexcel==2.5.14
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandoc==2.4
pandocfilters==1.5.1
parso==0.8.4
pdf2docx==0.5.8
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
plumbum==1.9.0
ply==3.11
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.1
PyMuPDF==1.26.4
pypandoc_binary==1.15
pyparsing==3.2.1
pypdf==6.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pyserial==3.5
PySoundFile==0.9.0.post1
python-calamine==0.4.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-json-logger==3.2.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
reedsolo==1.7.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.22.3
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
Send2Trash==1.8.3
setuptools==70.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.2
stack-data==0.6.3
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
whitenoise==6.8.2
wrapt==1.17.2
